service/requirements.txt
attrs==16.2.0
click==6.6
coreapi==2.0.8
Django==1.10.2
django-munin==0.2.1
django-rest-raml==0.1.0
djangorestframework==3.5.0
gunicorn==19.6.0
itypes==1.1.0
jsonref==0.1
Markdown==2.6.7
markdown2==2.3.1
psycopg2==2.6.2
PyYAML==3.12
raml-codec==0.1.2
ramlfications==0.1.9
requests==2.11.1
six==1.10.0
termcolor==1.1.0
Unipath==1.1
uritemplate==3.0.0
xmltodict==0.10.2
